+++ src/main/java/com/test/repository/BrandRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
 import com.test.entity.Brand;
 
 @Repository
-public interface BrandRepository extends JpaRepository<Brand, Integer>, JpaSpecificationExecutor<Brand> {
+public interface BrandRepository extends JpaRepository<Brand, Long>, JpaSpecificationExecutor<Brand> {
 	List<Brand> findByNameContainingIgnoreCase(String name);
 }
