+++ src/main/java/com/test/repository/ModelRepository.java
@@ -8,8 +8,8 @@ import com.test.entity.Model;
 
 
 
-public interface ModelRepository extends JpaRepository<Model,Integer> {
-	List<Model> findByBrandId(Integer brandId);
+public interface ModelRepository extends JpaRepository<Model,Long> {
+	List<Model> findByBrandId(Long brandId);
 	
 	
 }
